refactor: setup cogshop v1 model
