+++ cognite/powerops/resync/models/_shared_v1_v2/__init__.py
@@ -1,0 +1,3 @@
+"""
+This package contains helpers that are used by both the v1 and v2 models.
+"""
