+++ cognite/powerops/resync/models/base/__init__.py
@@ -4,7 +4,7 @@ methods for working with the models, and also provide a standardized interface f
 """
 
 from .asset_model import AssetModel
-from .asset_type import AssetType, NonAssetType
+from .asset_type import AssetType, NonAssetType, T_Asset_Type
 from .data_model import DataModel
 from .model import Model, T_Model
 from .cdf_resources import CDFSequence, CDFFile, CDFResource
@@ -13,6 +13,7 @@ from .graph_ql import PowerOpsGraphQLModel
 __all__ = [
     "AssetModel",
     "AssetType",
+    "T_Asset_Type",
     "NonAssetType",
     "DataModel",
     "Model",
